ai-course-code/requirements.txt
# AI大模型应用开发 - 课程依赖
# Python 3.10+ 推荐

# ============================================
# 基础依赖
# ============================================
python-dotenv>=1.0.0          # 环境变量管理
requests>=2.31.0              # HTTP请求
httpx>=0.25.0                 # 异步HTTP
pydantic>=2.5.0               # 数据验证
typing-extensions>=4.8.0      # 类型提示扩展

# ============================================
# OpenAI & LLM
# ============================================
openai>=1.6.0                 # OpenAI API
tiktoken>=0.5.0               # Token计数
tenacity>=8.2.0               # 重试机制

# ============================================
# LangChain 生态
# ============================================
langchain>=0.1.0              # LangChain核心
langchain-openai>=0.0.2       # OpenAI集成
langchain-community>=0.0.10   # 社区集成
langchain-core>=0.1.0         # 核心抽象
langgraph>=0.0.20             # LangGraph框架
langsmith>=0.0.75             # LangSmith调试

# ============================================
# 向量数据库
# ============================================
chromadb>=0.4.18              # Chroma向量数据库
faiss-cpu>=1.7.4              # FAISS (CPU版本)
# faiss-gpu>=1.7.4            # FAISS (GPU版本，可选)

# ============================================
# Embedding
# ============================================
sentence-transformers>=2.2.0  # 本地Embedding模型
# 中文Embedding模型会在课程中单独安装

# ============================================
# 文档处理
# ============================================
pypdf>=3.17.0                 # PDF解析
python-docx>=1.1.0            # Word文档
openpyxl>=3.1.2               # Excel文档
unstructured>=0.10.0          # 通用文档解析
markdown>=3.5.0               # Markdown解析
beautifulsoup4>=4.12.0        # HTML解析

# ============================================
# 知识图谱
# ============================================
neo4j>=5.14.0                 # Neo4j驱动
py2neo>=2021.2.3              # Neo4j Python工具

# ============================================
# Web框架
# ============================================
fastapi>=0.104.0              # FastAPI
uvicorn>=0.24.0               # ASGI服务器
python-multipart>=0.0.6       # 文件上传

# ============================================
# 数据处理
# ============================================
numpy>=1.24.0                 # 数值计算
pandas>=2.1.0                 # 数据分析

# ============================================
# Jupyter
# ============================================
jupyter>=1.0.0                # Jupyter Notebook
ipykernel>=6.27.0             # Jupyter内核
ipywidgets>=8.1.0             # Jupyter小部件

# ============================================
# 测试与调试
# ============================================
pytest>=7.4.0                 # 测试框架
pytest-asyncio>=0.21.0        # 异步测试

# ============================================
# 微调相关（第15-18章，可选安装）
# ============================================
# torch>=2.1.0                # PyTorch
# transformers>=4.35.0        # Transformers
# peft>=0.6.0                 # PEFT库
# accelerate>=0.24.0          # 加速库
# bitsandbytes>=0.41.0        # 量化支持
# datasets>=2.15.0            # 数据集
# trl>=0.7.0                  # TRL训练
# deepspeed>=0.12.0           # DeepSpeed

# ============================================
# 部署相关（第22章，可选安装）
# ============================================
# docker>=6.1.0               # Docker SDK
# redis>=5.0.0                # Redis
# prometheus-client>=0.19.0   # 监控

